pyproject.toml
[tool.poetry]
name = "ai-doc-composer"
version = "1.0.0"
description = "Automated documentary generation pipeline using orchestrated AI models for video captioning, narrative planning, speech synthesis, and video rendering."
authors = ["Andrii K. <ak111@student.london.ac.uk>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/andriikyselov/ai-doc-composer"
repository = "https://github.com/andriikyselov/ai-doc-composer"
keywords = ["ai", "documentary", "video-processing", "computer-vision", "nlp", "tts", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]
packages = [{ include = "ai_doc_composer", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12"
rich = "^13.7"
pydantic = "^2.6"
numpy = "1.24.4"
pillow = "^10.3"
accelerate = "^0.28"
transformers = "4.35.2"
ffmpeg-python = "^0.2"
tqdm = "^4.66"
google-generativeai = "^0.8.0"
python-dotenv = "^1.0"
soundfile = "^0.13.1"
openai-whisper = "^20231117"
jiwer = "^3.0.3"
gradio = "^5.0"
[tool.poetry.group.dev.dependencies]
pytest = "^8.1"

[tool.poetry.group.vibevoice]
optional = true

[tool.poetry.group.vibevoice.dependencies]
safetensors = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"